+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added `open_links` to `Markdown` and `MarkdownViewer` widgets
 - Added `background-tint` CSS rule https://github.com/Textualize/textual/pull/5117
 - Added `App.DEFAULT_MODE`
+- Added `Containers.HorizontalGroup` and `Containers.VerticalGroup`
 
 ### Fixed